requirements.txt
selenium==4.18.1
requests==2.31.0
pytest==8.0.0
allure-pytest==2.13.2
python-dotenv==1.2.1
ruff==0.9.6